+++ R/meta.normal.R
@@ -1,9 +1,7 @@
-# TODO make sure this code works when there is no natural variance !!!!!!!!!!!!
-
 # population-level parameter estimates for normally distributed parameters and parameter uncertainties
-# VARS Boolean denotes whether or not there is natural variance
 # MEANS Boolean denotes whether or not there is a non-zero mean
-meta.normal <- function(MU,SIGMA,MEANS=TRUE,VARS=TRUE,diagonal=FALSE,isotropic=FALSE,debias=TRUE,weights=NULL,precision=1/2)
+# VARS Boolean denotes whether or not there is natural variance-covariance
+meta.normal <- function(MU,SIGMA,MEANS=TRUE,VARS=TRUE,isotropic=FALSE,GUESS=NULL,debias=TRUE,weights=NULL,precision=1/2)
 {
   tol <- .Machine$double.eps^precision
   REML <- debias
@@ -28,10 +26,12 @@ meta.normal <- function(MU,SIGMA,MEANS=TRUE,VARS=TRUE,diagonal=FALSE,isotropic=F
   else
   { weights <- weights/mean(weights) }
 
-  # do we give variance to each dimension
-  VARS <- array(VARS,DIM)
   # do we give a mean to each dimension
   MEANS <- array(MEANS,DIM)
+  # do we give a variance to each dimension after J-transformation
+  if(length(VARS)==1) { VARS <- array(VARS,DIM) }
+  if(length(dim(VARS))<2) { VARS <- outer(VARS,FUN="&") }
+  if(isotropic) { VARS <- diag(diag(VARS)) }
 
   SCALE <- apply(MU,2,stats::mad)
   if(isotropic) { SCALE[] <- stats::median(SCALE) }
@@ -43,34 +43,30 @@ meta.normal <- function(MU,SIGMA,MEANS=TRUE,VARS=TRUE,diagonal=FALSE,isotropic=F
     else
     { SCALE[ZERO] <- 1 } # do no rescaling
   }
-  SCALE2 <- SCALE^2
 
   # observations
   OBS <- array(TRUE,c(N,DIM))
-  for(i in 1:N) { OBS[i,] <- diag(cbind(SIGMA[i,,])) < SCALE2/.Machine$double.eps }
+  for(i in 1:N) { OBS[i,] <- diag(cbind(SIGMA[i,,])) < SCALE^2/.Machine$double.eps }
   # zero out Inf-VAR observations, in case of extreme point estimates
   MU[!OBS] <- 0
   # zero out Inf-VAR correlations, in case of extreme point estimates
   for(i in 1:N) { for(j in which(!OBS[i,])) { SIGMA[i,j,-j] <- SIGMA[i,-j,j] <- 0 } }
 
   NOBS <- colSums(OBS) # number of observations per parameter
-  SUB <- NOBS<=1 # can't calculate variance for these
-  if(any(SUB)) { VARS[SUB] <- FALSE }
   SUB <- NOBS==0 # can't calculate mean for these
   if(any(SUB)) { MEANS[SUB] <- FALSE }
-
-  ZEROV <- !VARS
-  ZEROM <- !MEANS
+  SUB <- NOBS<=1 # can't calculate variance for these
+  if(any(SUB)) { VARS[SUB,] <- VARS[,SUB] <- FALSE }
+  vars <- diag(VARS)
 
   ####################
   # robust rescaling in case of outliers with ill-conditioned COV matrices
   SHIFT <- apply(MU,2,stats::median)
-  if(any(ZEROM)) { SHIFT[ZEROM] <- 0 }
+  if(any(!MEANS)) { SHIFT[!MEANS] <- 0 }
   # apply shift
   MU <- t( t(MU) - SHIFT )
   # initial guess of mu
-  mu <- SHIFT
-  mu[] <- 0
+  mu <- array(0,DIM)
 
   # apply scale
   MU <- t( t(MU)/SCALE )
@@ -80,59 +76,117 @@ meta.normal <- function(MU,SIGMA,MEANS=TRUE,VARS=TRUE,diagonal=FALSE,isotropic=F
   SIGMA <- SIGMA/SCALE
   SIGMA <- aperm(SIGMA,c(2,3,1)) # [id,x,y]
   # initial guess of sigma
-  sigma <- diag(1,nrow=length(mu))
+  sigma <- diag(diag(VARS))
 
   COV.mu <- sigma/N
-  if(any(ZEROV)) { sigma[ZEROV,] <- sigma[,ZEROV] <- 0 }
-  if(any(ZEROM)) { COV.mu[ZEROM,] <- COV.mu[,ZEROM] <- 0 }
+  if(any(!MEANS)) { COV.mu[!MEANS,] <- COV.mu[,!MEANS] <- 0 }
 
+  # potential unique sigma parameters
+  TRI <- upper.tri(sigma,diag=TRUE)
   # non-zero unique sigma parameters
-  UP <- upper.tri(sigma,diag=FALSE)
-  UP[ZEROV,] <- FALSE
-  UP[,ZEROV] <- FALSE
+  DUP <- TRI & VARS
 
-  DUP <- upper.tri(sigma,diag=TRUE)
-  DUP[ZEROV,] <- FALSE
-  DUP[,ZEROV] <- FALSE
+  # extract parameters from sigma matrix
+  COR <- TRUE # use correlations rather than covariances
+  sigma2par <- function(sigma)
+  {
+    if(isotropic)
+    {
+      par <- mean(sigma[VARS])
+      par <- nant(par,0) # no VARS
+    }
+    else
+    {
+      if(COR)
+      {
+        D <- diag(sigma)
+        d <- sqrt(D)
+        d[!vars | d<=.Machine$double.eps] <- 1
+        d <- d %o% d
+        sigma <- sigma/d
+        diag(sigma) <- D
+      }
+
+      par <- sigma[DUP]
+    }
+
+    return(par)
+  }
+
+  # construct sigma matrix from parameters
+  par2sigma <- function(par)
+  {
+    sigma <- diag(0,DIM)
+
+    sigma[DUP] <- par
+    sigma <- t(sigma)
+    sigma[DUP] <- par
+
+    if(!isotropic && COR)
+    {
+      D <- diag(sigma)
+      d <- sqrt(abs(D))
+      # d[d<=.Machine$double.eps] <- 1
+      d <- d %o% d
+      sigma <- sigma*d
+      diag(sigma) <- D
+    }
+
+    return(sigma)
+  }
 
   INF <- list(loglike=-Inf,mu=mu,COV.mu=sigma,sigma=sigma,sigma.old=sigma)
+  if(!is.null(GUESS))
+  {
+    GUESS$mu <- (GUESS$mu-SHIFT)/SCALE
+    GUESS$sigma <- t(GUESS$sigma/SCALE)/SCALE
+
+    INF$mu[MEANS] <- GUESS$mu[MEANS]
+    INF$sigma[VARS] <- GUESS$sigma[VARS]
+    sigma <- INF$sigma
+
+    INF$sigma.old <- INF$sigma
+    INF$COV.mu <- INF$sigma/N
+    if(any(!MEANS)) { INF$COV.mu[!MEANS,] <- INF$COV.mu[,!MEANS] <- 0 }
+  }
+
+  par <- sigma2par(sigma)
 
   # negative log-likelihood
   CONSTRAIN <- TRUE # constrain likelihood to positive definite
-  COR <- TRUE # use correlations rather than covariances
   nloglike <- function(par,REML=debias,verbose=FALSE,zero=0)
   {
-    zero <- -zero # now log-likelihood
+    if(!verbose) { INF <- Inf }
+    zero <- -zero # log-likelihood calculated below
     sigma <- par2sigma(par)
 
     # check for bad sigma matrices
     if(any(VARS))
     {
-      V <- abs(diag(sigma[VARS,VARS]))
+      # not sure how this happened once
+      if(any(abs(par)==Inf)) { return(INF) }
+
+      V <- abs(diag(sigma)[vars])
       V <- sqrt(V)
       # don't divide by zero
       TEST <- V<=.Machine$double.eps
       if(any(TEST)) { V[TEST] <- 1 }
       V <- V %o% V
-      S <- sigma[VARS,VARS]/V
+      S <- sigma[vars,vars]/V
       S <- eigen(S)
-      if(CONSTRAIN)
+      if(any(S$values<0))
       {
-        if(any(S$values<0))
+        if(CONSTRAIN)
+        { return(INF) }
+        else
         {
-          if(!verbose)
-          { return(Inf) }
-          else
-          { return(INF) }
+          S$values <- pmax(S$values,0)
+          S <- S$vectors %*% diag(S$values,length(S$values)) %*% t(S$vectors)
+          sigma[vars,vars] <- S * V
+          sigma[!VARS] <- 0
         }
-      }
-      else
-      {
-        S$values <- pmax(S$values,.Machine$double.eps)
-        S <- S$vectors %*% diag(S$values) %*% t(S$vectors)
-        sigma[VARS,VARS] <- S * V
-      }
-    }
+      } # end if negative variances
+    } # end bad sigma check
 
     # estimate mu exactly | sigma
     S <- P <- array(0,c(N,DIM,DIM))
@@ -152,9 +206,8 @@ meta.normal <- function(MU,SIGMA,MEANS=TRUE,VARS=TRUE,diagonal=FALSE,isotropic=F
     # if(any(ZEROM)) { mu[ZEROM] <- 0 } # should be okay
 
     # sum up log-likelihood
-    loglike <- REML/2*log(abs(det(COV.mu[MEANS,MEANS,drop=FALSE]))) # - DIM*(N-REML)/2*log(2*pi)
-    zero <- zero + DIM*(N-REML)/2*log(2*pi)
-    zero <- 2*zero # for 1/2 below
+    loglike <- -REML/2*PDlogdet(P.mu[MEANS,MEANS]) # - DIM*(N-REML)/2*log(2*pi)
+    zero <- 2/N*zero + DIM*(1-REML/N)*log(2*pi) # for below
     RHS <- 0
     LHS <- P.mu
     for(i in 1:N)
@@ -178,75 +231,18 @@ meta.normal <- function(MU,SIGMA,MEANS=TRUE,VARS=TRUE,diagonal=FALSE,isotropic=F
     # update sigma
     if(any(VARS))
     {
-      K <- sqrtm(sigma[VARS,VARS],pseudo=TRUE)
-      K <- K %*% PDfunc(LHS[VARS,VARS],function(m){1/sqrt(m)},pseudo=TRUE)
-      sigma[VARS,VARS] <- K %*% RHS[VARS,VARS] %*% t(K)
       # we are solving the unconstrained sigma above and then projecting back to the constrained sigma below
+      K <- sqrtm(sigma[vars,vars],pseudo=TRUE)
+      K <- K %*% PDfunc(LHS[vars,vars],function(m){1/sqrt(m)},pseudo=TRUE)
+      sigma[vars,vars] <- K %*% RHS[vars,vars] %*% t(K)
+
       # not sure how approximate this is, but will do numerical optimization afterwards
-      if(isotropic)
-      { sigma <- diag( mean(diag(sigma)), DIM ) }
-      else if(diagonal)
-      { sigma <- diag( diag(sigma) , DIM ) }
+      sigma[!VARS] <- 0
+      if(isotropic) { sigma[VARS] <- mean(sigma[VARS]) }
     }
 
     R <- list(loglike=loglike,mu=mu,COV.mu=COV.mu,sigma=sigma,sigma.old=sigma.old)
     return(R)
-  }
-
-  # extract parameters from sigma matrix
-  sigma2par <- function(sigma)
-  {
-    if(isotropic)
-    { par <- mean(diag(sigma)) }
-    else if(diagonal)
-    { par <- diag(sigma)[VARS] }
-    else
-    {
-      if(COR)
-      {
-        D <- diag(sigma)
-        d <- sqrt(D)
-        d[d<=.Machine$double.eps] <- 1
-        d <- d %o% d
-        sigma <- sigma/d
-        diag(sigma) <- D
-      }
-
-      par <- sigma[DUP]
-    }
-
-    return(par)
-  }
-
-  # construct sigma matrix from parameters
-  par2sigma <- function(par)
-  {
-    if(isotropic)
-    { sigma <- diag(par,DIM) }
-    else if(diagonal)
-    {
-      sigma <- diag(0,DIM)
-      diag(sigma)[VARS] <- par
-    }
-    else
-    {
-      sigma <- array(0,c(DIM,DIM))
-      sigma[DUP] <- par
-      sigma <- t(sigma)
-      sigma[DUP] <- par
-
-      if(COR)
-      {
-        D <- diag(sigma)
-        d <- sqrt(D)
-        d[d<=.Machine$double.eps] <- 1
-        d <- d %o% d
-        sigma <- sigma*d
-        diag(sigma) <- D
-      }
-    }
-
-    return(sigma)
   }
 
   ##############
@@ -297,9 +293,9 @@ meta.normal <- function(MU,SIGMA,MEANS=TRUE,VARS=TRUE,diagonal=FALSE,isotropic=F
     else # proceed as usual
     {
       # Standardized error
-      ERROR <- (NSOL$sigma - SOL$sigma)[VARS,VARS] # absolute error
+      ERROR <- (NSOL$sigma - SOL$sigma)[vars,vars] # absolute error
       ERROR <- ERROR %*% ERROR # square to make positive
-      K <- PDsolve(NSOL$sigma[VARS,VARS],pseudo=TRUE)
+      K <- PDsolve(NSOL$sigma[vars,vars],pseudo=TRUE)
       ERROR <- K %*% ERROR %*% K # standardize to make ~1 unitless
       ERROR <- sum(abs(diag(ERROR)),na.rm=TRUE)
       count.0 <- 0
@@ -331,17 +327,11 @@ meta.normal <- function(MU,SIGMA,MEANS=TRUE,VARS=TRUE,diagonal=FALSE,isotropic=F
 
     if(isotropic)
     {
-      # in case sigma is zero
-      if(!known) { MIN <- mean(diag(COV.mu)) }
       lower <<- 0
       upper <<- Inf
-    }
-    else if(diagonal)
-    {
+
       # in case sigma is zero
-      if(!known) { MIN <- diag(COV.mu)[VARS] }
-      lower <<- 0
-      upper <<- Inf
+      if(!known) { MIN <- mean(COV.mu[VARS]) }
     }
     else
     {
@@ -396,7 +386,6 @@ meta.normal <- function(MU,SIGMA,MEANS=TRUE,VARS=TRUE,diagonal=FALSE,isotropic=F
   set.parscale()
 
   # not sure if the iterative solution always works in constrained problems
-  # if(any(ZEROV) || any(ZEROM))
   if(any(VARS))
   {
     # liberal parscale
@@ -421,23 +410,16 @@ meta.normal <- function(MU,SIGMA,MEANS=TRUE,VARS=TRUE,diagonal=FALSE,isotropic=F
     set.parscale(TRUE) # more accurate parscale for numderiv
     DIFF <- genD(par,nloglike,parscale=parscale,lower=lower,upper=Inf)
     COV.sigma <- cov.loglike(DIFF$hessian,DIFF$gradient)
+    # HESS <- DIFF$hessian - outer(DIFF$gradient) # Hessian penalized by non-zero gradient
   }
   else
-  {
-    if(isotropic)
-    { P <- 1 }
-    else if(diagonal)
-    { P <- DIM }
-    else
-    { P <- DIM*(DIM+1)/2 }
-    COV.sigma <- matrix(0,P,P)
-  }
+  { COV.sigma <- diag(0,DIM*(DIM+1)/2) }
 
   loglike <- -nloglike(par,REML=FALSE) # non-REML for AIC/BIC
 
   # rescale
   loglike <- loglike - log(prod(SCALE))
-  mu <- mu*SCALE + SHIFT
+  mu <- SCALE*mu + SHIFT
   sigma <- t(sigma*SCALE)*SCALE
   COV.mu <- t(COV.mu*SCALE)*SCALE
   if(any(VARS))
@@ -445,11 +427,11 @@ meta.normal <- function(MU,SIGMA,MEANS=TRUE,VARS=TRUE,diagonal=FALSE,isotropic=F
     SCALE <- SCALE %o% SCALE
     if(isotropic)
     { SCALE <- SCALE[1] }
-    else if(diagonal)
-    { SCALE <- diag(SCALE) }
     else
     { SCALE <- SCALE[DUP] }
+
     COV.sigma <- t(COV.sigma*SCALE)*SCALE
+    # HESS <- t(HESS/SCALE)/SCALE
   }
 
   # AIC
@@ -457,64 +439,34 @@ meta.normal <- function(MU,SIGMA,MEANS=TRUE,VARS=TRUE,diagonal=FALSE,isotropic=F
   q <- DIM
   qk <- sum(MEANS)
   if(isotropic)
-  { nu <- min(1,sum(VARS)) }
-  else if(diagonal)
-  { nu <- sum(VARS) }
+  { nu <- max(VARS) }
   else
   { nu <- sum(DUP) }
   K <- qk + nu
 
   AIC <- 2*K - 2*loglike
-  if(nu==0) # no variance parameters estimated, no bias
-  { AICc <- AIC }
-  else # some variance parameters estimated
-  {
-    AICc <- (q*n-qk)*2*K/max(q*n-K-nu,0) - 2*loglike
-    AICc <- nant(AICc,Inf)
-  }
+  AICc <- (q*n-qk)/max(q*n-K-nu,0)*2*K - 2*loglike
+  AICc <- nant(AICc,Inf)
   BIC <- K*log(N) - 2*loglike
 
   names(mu) <- NAMES
   dimnames(COV.mu) <- list(NAMES,NAMES)
   dimnames(sigma) <- list(NAMES,NAMES)
-  # sigma <- sigma[VARS,VARS]
-  if(isotropic)
-  {
-    NAMES2 <- "sigma-sigma"
-    dimnames(COV.sigma) <- list(NAMES2,NAMES2)
-  }
-  else if(diagonal)
-  {
-    # basic structure
-    NAMES2 <- NAMES[VARS]
-    NAMES2 <- paste0(NAMES2,"-",NAMES2)
-    dimnames(COV.sigma) <- list(NAMES2,NAMES2)
+  dimnames(VARS) <- list(NAMES,NAMES)
 
-    # promote to general structure below
-    NAMES3 <- outer(NAMES,NAMES,function(x,y){paste0(x,"-",y)})
-    if(any(VARS))
-    { NAMES3 <- NAMES3[DUP] }
-    else
-    { NAMES3 <- NAMES3[upper.tri(NAMES3,diag=TRUE)] }
-    # zeroed COV matrix
-    TEMP <- matrix(0,length(NAMES3),length(NAMES3))
-    dimnames(TEMP) <- list(NAMES3,NAMES3)
-    # copy over non-zero COV
-    TEMP[NAMES2,NAMES2] <- COV.sigma
-    COV.sigma <- TEMP
-  }
-  else
-  {
-    NAMES2 <- outer(NAMES,NAMES,function(x,y){paste0(x,"-",y)})
-    if(any(VARS))
-    { NAMES2 <- NAMES2[DUP] }
-    else
-    { NAMES2 <- NAMES2[upper.tri(NAMES2,diag=TRUE)] }
-    dimnames(COV.sigma) <- list(NAMES2,NAMES2)
-  }
+  # full matrix to copy into - all possible matrix elements
+  NAMES2 <- outer(NAMES,NAMES,function(x,y){paste0(x,"-",y)})
+  COV.SIGMA <- diag(0,DIM^2)
+  dimnames(COV.SIGMA) <- list(NAMES2,NAMES2)
+  # copy over non-zero elements
+  COV.SIGMA[which(DUP),which(DUP)] <- COV.sigma
+  # unique matrix elements only
+  COV.sigma <- COV.SIGMA[which(TRI),which(TRI)]
+  # dimnames(HESS) <- list(NAMES2,NAMES2)
 
   R <- list(mu=mu,sigma=sigma,COV.mu=COV.mu,COV.sigma=COV.sigma,loglike=loglike,AIC=AIC,AICc=AICc,BIC=BIC,isotropic=isotropic)
   R$VARS <- VARS # need to pass this if variance was turned off due to lack of data
   R$MEANS <- MEANS
+  # R$HESS <- HESS
   return(R)
 }